Fix for trainer to match original implementation
In original implementation and the setup we tested, if separate_funcs
was True, generator received different batch of random latents than
discriminator to optimize for and from now on it does in this
implementation too. Also, fixing a bug in discriminator training in
which if D_training_repeats > 1, discriminator optimized for the same
real images and fake latents for all iterations, which doesn't really
make sense.