Add empty close method as default for Dataset
Datasets implemented in the future might want to do some cleanup when
they are destroyed (one example is OldH5Dataset, which closes the
underlying h5 file). To avoid AttributeError in train.py, all datasets
now are closed at the end of training. This should not break the already
weird behaviour of crashing at the end of training for the already
existing saved datasets.